requirements.txt


```txt
ultralytics
streamlit
opencv-python
numpy
pandas
Pillow